운영계 접속정보 수정 반영(20231019)

+++ src/main/java/com/zinidata/eatout/vo/output/AgileMarketMainOutVO.java
@@ -10,7 +10,7 @@ public class AgileMarketMainOutVO {
     private String  upjong3Cd;
     private String  megaNm;
     private String  megaCd;
-    private int     calcSaleAmt;
+    private double     calcSaleAmt;
     private double  salePer;
     private double  storePer;
     private double  calcPer;
